requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-modelcluster==5.0.1
django-storages==1.9.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
google-api-core==1.17.0
google-auth==1.16.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
l18n==2018.5
Pillow==7.1.2
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9
webencodings==0.5.1
whitenoise==5.1.0
Willow==1.3
XlsxWriter==1.2.8
